Separate solver logic from algorithm
